chore: extract outlined button as component
