+++ sample/stories/ProfileCard.stories.tsx
@@ -1,7 +1,7 @@
 import type { ComponentMeta, ComponentStory } from '@storybook/react-native';
 import React from 'react';
 
-import { ProfileCard as ProfileCardComponent } from '@sendbird/uikit-react-native-foundation';
+import { OutlinedButton, ProfileCard as ProfileCardComponent } from '@sendbird/uikit-react-native-foundation';
 
 const ProfileCardMeta: ComponentMeta<typeof ProfileCardComponent> = {
   title: 'ProfileCard',
@@ -14,6 +14,10 @@ const ProfileCardMeta: ComponentMeta<typeof ProfileCardComponent> = {
     username: {
       name: 'Username',
       control: { type: 'text' },
+    },
+    button: {
+      name: 'Button',
+      control: { type: null },
     },
     bodyLabel: {
       name: 'Body label',
@@ -36,3 +40,6 @@ export default ProfileCardMeta;
 
 type ProfileCardStory = ComponentStory<typeof ProfileCardComponent>;
 export const Default: ProfileCardStory = (args) => <ProfileCardComponent {...args} />;
+export const WithButton: ProfileCardStory = (args) => (
+  <ProfileCardComponent {...args} button={<OutlinedButton>{'Message'}</OutlinedButton>} />
+);
